Prepare for formatting
