+++ .github/workflows/clang-format-check.yml
@@ -1,0 +1,25 @@
+name: clang-format Check
+on: [push, pull_request]
+jobs:
+  formatting-check:
+    name: Formatting Check
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        path:
+          - check: 'src'
+            exclude: '\.proto$'
+          - check: 'include'
+          - check: 'nmt'
+          - check: 'python'
+          - check: 'scripts'
+          - check: 'tests'
+          - check: 'examples'
+    steps:
+    - uses: actions/checkout@v2
+    - name: Run clang-format style check for C/C++/Protobuf programs.
+      uses: jidicula/clang-format-action@v4.8.0
+      with:
+        clang-format-version: '13'
+        check-path: ${{ matrix.path['check'] }}
+        exclude-regex: ${{ matrix.path['exclude'] }}
