+++ CMakeLists.txt
@@ -105,6 +105,9 @@ include(json)
 # variant
 include(variant)
 
+# optional
+include(optional)
+
 if(FF_USE_PYTHON)
   list(APPEND FF_CC_FLAGS
     -DBINDINGS_AUGMENT_PYTHONPATH)
@@ -199,7 +202,7 @@ else()
 endif()
 
 target_include_directories(flexflow PUBLIC ${FLEXFLOW_INCLUDE_DIRS})
-target_link_libraries(flexflow ${LEGION_LIBRARY} nlohmann_json::nlohmann_json mpark_variant)
+target_link_libraries(flexflow ${LEGION_LIBRARY} nlohmann_json::nlohmann_json mpark_variant optional)
 
 if(FF_USE_NCCL)
   add_dependencies(flexflow ${NCCL_NAME})
