+++ include/flexflow/dominators.h
@@ -4,7 +4,7 @@
 #include "flexflow/basic_graph.h"
 #include "flexflow/graph_structures.h"
 #include "flexflow/utils/dot/dot_file.h"
-#include "tl/optional.h"
+#include "tl/optional.hpp"
 #include <queue>
 #include <algorithm>
 #include <functional>
