+++ include/flexflow/ops/pool_2d_params.h
@@ -1,0 +1,35 @@
+#ifndef _FLEXFLOW_POOL_2D_PARAMS_H
+#define _FLEXFLOW_POOL_2D_PARAMS_H
+
+#include "flexflow/ffconst.h"
+#include "flexflow/parallel_tensor.h"
+
+namespace FlexFlow {
+
+struct Pool2DParams {
+  int kernel_h, kernel_w, stride_h, stride_w, padding_h, padding_w;
+  PoolType pool_type;
+  ActiMode activation;
+
+  bool is_valid(ParallelTensorShape const &input) const;
+  void solve_dims(ParallelTensorShape const &input, 
+                  ParallelDim output_dims[MAX_TENSOR_DIM], int* output_ndims) const;
+private:
+  int output_size(ParallelTensorShape const &input,
+                  ParallelDim output_dims[MAX_TENSOR_DIM]) const; 
+};
+
+bool operator==(const Pool2DParams&, const Pool2DParams&);
+
+
+} // namespace FlexFlow
+
+namespace std {
+  template <>
+  struct hash<FlexFlow::Pool2DParams> {
+    size_t operator()(const FlexFlow::Pool2DParams&) const;
+  };
+} // namespace std
+
+
+#endif // _FLEXFLOW_POOL_2D_PARAMS_H
