+++ include/flexflow/utils/dot/dot_file.h
@@ -3,7 +3,7 @@
 
 #include <fstream>
 #include <sstream>
-#include "tl/optional.h"
+#include "tl/optional.hpp"
 #include <vector>
 #include <cassert>
 #include <map>
